Create question1.cpp
The code implements the Dijkstra and Graph header files from the repository, maing use of the variabole and declarations already provided in the header files. The code handles errors while opening and closing the DIMACS file downloaded and included for manual parsing. Precise timing and typ deduction and range-based for loop. The code also assumes custom binary format if and as applicable.